+++ lib/dart_jsonwebtoken.dart
@@ -1,6 +1,6 @@
 library jsonwebtoken;
 
 export 'src/jwt.dart';
-export 'src/errors.dart';
+export 'src/exceptions.dart';
 export 'src/algorithms.dart';
 export 'src/keys.dart';